backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
